+++ mobile/src/components/Buttons/Floating/index.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import {StyleSheet} from 'react-native';
 import {RectButtonProperties} from 'react-native-gesture-handler';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 
@@ -9,6 +10,10 @@ interface Props extends ContainerProps, WrapperProps, RectButtonProperties {
   iconColor: string;
   iconSize?: number;
 }
+
+const styles = StyleSheet.create({
+  elevated: {elevation: 8},
+});
 
 const Floating: React.FC<Props> = ({
   backgroundColor,
@@ -21,7 +26,7 @@ const Floating: React.FC<Props> = ({
 }) => {
   return (
     <Wrapper {...{position}}>
-      <Container {...{size, backgroundColor, ...rest}}>
+      <Container {...{size, backgroundColor, ...rest}} style={styles.elevated}>
         <Icon {...{name, color}} size={iconSize} />
       </Container>
     </Wrapper>
